[feature] Replace words by emojis for signing phrase
Note to developers:

If you see white rectangles instead of some emojis it is because your
font doesn't support them, but you should still be able to edit
the file in a text editor.
If your text editor corrupts the information in the emoji's file when editing
it, use another one.