[#4181] use react/text to show sign message text
Right now the sign message text is shown using the `amount-input` component.
This component uses a multi-line `react/input-text` to show the message, but the
input is always disabled and the text is shown unmodified
(src/status_im/ui/screens/wallet/send/views.cljs:289).  Multi-line input-texts
do not expand their height based on the content, so the message is cut off when
displayed.

This commit replaces the use of the `amount-input` component by a simple
`react/text` with `selectable` property set to true. The behavior should be
almost the same than before, but now the all the content is visible.
(NOTE that the text is already under a scroll view).