Refactor: move global information out of local fn
Here we are moving knowledge of the greater world (the shap of app-db) out of a local helper function. We push this information up.

We don't want to pass more info than is needed to a fn as then it becomes hard to tell what it actually needs.

Make notes about the types, and added pre check to make sure that type assumptions line up during refactorgin.